change joueur.scala

+++ Version_0/src/controleur/Coordonnees2D.scala
@@ -1,6 +1,6 @@
 package controleur
 
-class Coordonnees2D ( x: Double, y: Double)extends AbstractCoordonnees(x:Double,y:Double){
+class Coordonnees2D ( x: Int, y: Int)extends AbstractCoordonnees(x:Int,y:Int){
    
   /*in order to compare 2 coordonnees*/
   def equals(c2: Coordonnees2D):Boolean ={
